Update files for merge

+++ src/rendering/DepthPrepass.cpp
@@ -175,7 +175,7 @@ void DepthPrepass::BuildDescriptors() {
     }
 
     // Camera Transform UBO
-    for (size_t i = 0; i < (size_t)vkutil::MAX_FRAMES_IN_FLIGHT; i++) {
+    for (size_t i = 0; i < vkutil::MAX_FRAMES_IN_FLIGHT; i++) {
         VkDescriptorBufferInfo bufferInfo{};
         bufferInfo.buffer = camera->GetBuffers()[i].buffer;
         bufferInfo.offset = 0;
